Added a Timer class. This is real slick - you basically add timers by name and specify a time to wait in milliseconds. Then you can periodically check up on them to either see how much time is left or if they're done.
Here's an example:

timer.add("hi", 1000);
timer.add("wahoo",3000);
timer.add("test", 5000);

bool test = false;
while(1) {
	if(timer.isDone("hi") && !test) {
		cout << "hi done at" << getTime();
		timer.remove("wahoo");
		test = true;
	}
	if(timer.isDone("wahoo")) {
		cout << "Wahoo done at" << getTime();
	}
	if(timer.isDone("test")) {
		cout << "test done at" << getTime();
		exit(0);
	}
}

+++ utilities.hpp
@@ -253,7 +253,7 @@ long long getMilliTime()
 	return retVal;
 }
 
-int getElapsedTime(long long startTime) {
+long long getElapsedTime(long long startTime) {
 	long long endTime = getMilliTime();
 	return endTime - startTime;
 }
